Fix ThemeButton state desync
Maintaining an "enabled" state is redundant as we can derive that from the theme variable from the useTheme hook. However, we can not set this until the page has fully loaded, as client cached data is not available until then. To fix this, render an empty slider until data is available.

It should not flicker in/out of existence when switching pages as it is part of the root layout, but there seems to be an issue with the Navbar not maintaining state. Will investigate, but out of the scope of this PR.