Task creation no longer broken
Categories selector connects to database and allows selection of a category.

Note that I used id=0 as the "Other" category; this is somewhat of a hack. Also, there are two extremely poor usages of useEffect, but I don't want to deal with Headless and caching so they're going to stay in.